busxando problema de boton de pruebas

+++ callbacks.c
@@ -123,17 +123,21 @@ void callbackTest( GtkWidget *widget, gpointer   data){
    line.last_accessed=3;
    line.times_accessed=4;
    line.first_accessed=5;
-   line.content=(long[]){1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
+   line.content=(unsigned[]){1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
    void * user_use;
    writeLineCacheInstructions(0, &line, 7);
-   line.content=(long[]){1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
+/*
+   line.content=(unsigned[]){1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
    writeLineCacheData(0, &line, 7);
-   line.content=(long[]){1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
+   line.content=(unsigned[]){1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
    writeLineCache(1, &line, 7);
+
    showCacheLineInstructions(0, 7);
+*/
    showCacheLineData(0, 7);
    showCacheLine(1, 7);
    showCacheLine(1, 6);
+
    setMemoryAddressColor(134217732, 5);
    setColorDataCacheLine(0, 2, 9);
    setColorInstructionsCacheLine(0, 2, 10);
